Update README.md
Technology is evolving, and the adoption of digital technologies is rapidly increasing; so are the Cyber and Information security risks. This Strategy sets out, at a high level, the Hays approach to the response and mitigation of Cyber and Information risks facing the business as far as possible. It guides the Business and IT to assist them in making well-informed risk-based decisions for 34 countries in Hays world.

This Strategy is designed to support efforts to Detect, React to and Defend against Cyber and Information threats and risks, highlighting the organisational process and technological responses needed to support the Strategy at both a global and regional level. A Security awareness program is in place at Hays globally and runs alongside this Strategy. The primary audience of this document is the IT organisation and Cyber Security / Risk Management roles or functions in Hays world for all countries under Hays Group Enterprise.

Disclaimer: This document is publicly available for the short version (An Executive copy only) of the Hays Global Cyber & InfoSec Strategy upon request if anyone like to do business at Hays (ideally vendors and the 3rd party supporting company). The whole Strategy can not be available as per GDPR 1998 & Act of 2018.